滚动训练框架搭建完成

+++ mc_training/dataset/mc_dataset.py
@@ -108,7 +108,7 @@ class RollingDataset(Dataset):
         # 添加单通道维度，确保形状为 [1, window_size, feature_dim]
         x = x.unsqueeze(0)
 
-        return x, self.img_data[idx], torch.tensor(y, dtype=torch.long)
+        return x, self.img_data[idx].to(torch.float32), torch.tensor(y, dtype=torch.long)
 
 
 # Example usage
